+++ src/main/java/org/fife/ui/rsyntaxtextarea/WrappedSyntaxView.java
@@ -256,7 +256,7 @@ return p + 1;
 		if (host.getEOLMarkersVisible()) {
 			g.setColor(host.getForegroundForTokenType(Token.WHITESPACE));
 			g.setFont(host.getFontForTokenType(Token.WHITESPACE));
-			g.drawString("\u00B6", x, y-fontHeight);
+			g.drawString("\u00B6", x, (float) y-fontHeight);
 		}
 
 	}
@@ -454,7 +454,7 @@ return p + 1;
 		if (host.getEOLMarkersVisible()) {
 			g.setColor(host.getForegroundForTokenType(Token.WHITESPACE));
 			g.setFont(host.getFontForTokenType(Token.WHITESPACE));
-			g.drawString("\u00B6", x, y-fontHeight);
+			g.drawString("\u00B6", x, (float) y-fontHeight);
 		}
 
 	}
@@ -840,7 +840,7 @@ return p + 1;
 		if (tabSize == 0)
 			return x;
 		int ntabs = ((int) x - tabBase) / tabSize;
-		return tabBase + ((ntabs + 1) * tabSize);
+		return tabBase + ((ntabs + 1f) * tabSize);
 	}
